icons and edit

+++ actions/delete-employee.php
@@ -2,7 +2,7 @@
     require_once '../includes/connection.php';
     // Verificar si el empleado está autenticado
     if (!isset($_SESSION['authenticated']) || !$_SESSION['authenticated']) {
-        header("Location: ../index.php"); // Si no está autenticado
+        header("Location: ../index.php");
         exit();
     }
 
@@ -19,17 +19,43 @@
     if(isset($_GET['dni'])) {
         $dni = $_GET['dni'];
 
-        // Eliminar el empleado
-        $query = "DELETE FROM empleados WHERE dni = '$dni'";
+        if (isset($_POST['confirm']) && $_POST['confirm'] == 'yes') {
+            // Eliminar el empleado
+            $query = "DELETE FROM empleados WHERE dni = '$dni'";
+            $result = mysqli_query($db, $query);
 
-        $result = mysqli_query($db, $query);
-
-        if($result) {
-            // ALERTA
-            header("Location: ../modules/employees.php");
-            exit(); 
-        } else {
-            echo "Hubo un error al eliminar el empleado."; // ALERTA
+            if($result) {
+                // ALERTA
+                header("Location: ../modules/employees.php");
+                exit(); 
+            } else {
+                echo "Hubo un error al eliminar el empleado."; // ALERTA
+            }
         }
-    } 
+
+        // Información del empleado
+        $query = "SELECT * FROM empleados WHERE dni = '$dni'";
+        $result = mysqli_query($db, $query);
+        $empleado = mysqli_fetch_assoc($result);
+    } else {
+        header("Location: ../modules/employees.php");
+        exit();
+    }
 ?>
+
+<?php include_once("../includes/header-admin.php"); ?>
+
+<main>
+    <h1>Eliminar Empleado</h1>
+    <div class="container">
+        <p>¿Seguro que quieres eliminar al empleado '<?php echo '<b>' . $empleado['nombre'] . ' ' . $empleado['apellidos'] . '</b>'; ?>'? </p>
+        <form action="delete-employee.php?dni=<?php echo $dni; ?>" method="POST">
+            <input type="hidden" name="confirm" value="yes">
+            <button class="btn btn-delete" type="submit">Eliminar</button>
+            <a href="../modules/employees.php" class="btn btn-cancel">Cancelar</a>
+        </form>
+    </div>
+</main>
+
+<?php include_once("../includes/footer.php"); ?>
+
